examples: split Data Pipeline (24) into multi-file project
Split monolithic main.py into three files:
- stages.py: stage agents with tools and simulated data
- pipeline.py: DelegationPlan DAG definition
- main.py: slim orchestrator entry point
